+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## 2025-07-31
 
-- Added docs, see [docs/README.md](docs/README.md)
+- Added docs, see [docs/README.md](docs/README.md).
+- Demoted aliases now don't show up in help at all.